Link to technology notes
Added link to technology notes, if it exists. It's provided along with a
link to user agent notes. However, it's subject to the current bug that
if user agent notes not present in the XML, the link not output; that
will hopefully be fixed separately. Alternatively, this link could be
added to the footer, but didn't do that for this commit. This closes
ACTION-188 https://www.w3.org/WAI/GL/track/actions/188.