Suggest using a more meaningful label
Minor suggestion to change the `aria-label` to be something that describes what the `<nav>` is for rather than the design pattern. The function of 'breadcrumbs' may not be clear to some people (or may not translate well... I'm guessing!) and how they are implemented (as either an indcation of ones location in the heirarchy or as an indication of ones journey through a website) is ambiguous.