Update user-interface-component-contrast-minimum.html
Removing "all" and "interactive" based on feedback from github comments and discussion at AGWG call.  Adding a note about the request to simplify this SC by removing any need to measure thickness and just going with a 3:1 ratio.  Adding "state"  to the note about what is an essential visual indicator.