Removed 'changes of content'
Like the bullet point on secondary content, this seems to be a different
matter than interruptions. A change of content is not a change of
context, and does not necessarily mean an interruption. By allowing a
user to postpone or suppress new content, the site may not even function
anymore.