Revised functionality as per meeting May 15th
Edited all requirements that imply that content needs to be fully tested in all orientation. SC is about orientation being locked, rather than ensuring that everything works in all orientations, which has compounding testing implications if it has to be tested after resizing SC is implemented, and reflow SC is implemented. This makes it a simple test. Turn scree orientation, does the page rotate?