removed a drag-n-drop from non-drag section.
this sentence:
"it does not use pre-defined gestures in the same way as the swiping or dragging of content or controls, or the drawing of specific shape patterns."

Included "dragging of content", which seems on a fresh read to mean drag & drop. I think it works better without that little snippet.