Implemented mbgover' change suggestions.
Removed dragging in general (but left in 'dragging nof a slider thumb'), incorporated a variant of what Michael had suggested as a replacement text. Changed the last example (floor-planning app)  away from the confusing drag-n-drop focus, and made it into a Kanban widget which allows horizontal swiping to move elements to adjacent silos.