Understanding 2.4.6: remove irrelevant benefits
The "People who have difficulty using their hands" benefit here is not explained, and appears to be more of a benefit assuming that headings are correctly marked up per 1.3.1, and that the users are either
navigating an automated TOC that links to headings, or are using additional AT that allows for heading-based navigation. All of this is not related to whether or not headings are actually descriptive.

The "This Success Criterion helps people who use screen readers" benefit again hinges more on 1.3.1 first and foremost. As it's not a requirement of THIS particular SC that headings be correctly marked up, it's
not really a benefit per se of this SC directly. May reintroduce this in reworded form.

The "This Success Criterion may also help users with low vision who can see only a few words at a time" benefit seems to be a complete non-sequitur, as descriptive headings/labels may vary in length and won't
necessarily benefit this user group in the way described.