Separated aspects of the existing techniques
G208 was too general. I removed all its examples which matched the label and added them to the G210 technique, which is about matching. I then made G208 be about including the text in the name. I also modified the Japanese version, which seemed to me to be innacurate and not helpful. Someone who speaks better Japanese than me should verify my rewording.