added placeholder in section 'boundaries'
In section "Boundaries" the example used is a button that does not necessarily need a contrasty outline in text acts as indicator and has  4.5:1 or better contrast. The same is probably true for placeholder (contrast above 4.5:1) before a text field gets focused. I have just added that case (we had questions around that).