Another take on Pointer Gestures
A re-write emphasizing the need for single-point activation, treating both directional swiping and dragging gestures as path-based gestures. Excluded are free-form drag-and-drop and free-form drawing (handwriting input, signature). I have rearranged examples and added a section with examples of alternatives for single-point activation to give implementers a clearer idea what the can do to meet this SC.