Update identify-purpose.html
Updating techniques. The technique listed for symbols does not exist in the COGA techniques document. It also is highly prescriptive -- _enabling_ symbols is an outcome of the author indicating purpose programmatically. So I have named a higher level general technique to cover this author action.
If there _is_ a specific additional task authors need to do to enable symbols, it needs to be better phrased so it is clear.