Add example of directional-only pointer gesture
While it's implied/is a subset of the existing first gesture (where points 1, 2 and 3 have to be in a straight line), having an explicit example here that calls these types of gestures out will make it more
immediately obvious that these are also covered. As this is  the simplest subset of the existing first figure, this was also made the first example (so the complexity builds up from the 1-to-2 straight line
example, to the 1-2-3 example).

Some added `<strong>` and `<em>` added to make it more immediately clear when something IS or ISN'T path-based/free-form.