Errata/clarification of 2.2.1 Timing Adjustable extension
Based on the discussion in https://lists.w3.org/Archives/Public/w3c-wai-gl/2020JanMar/0161.html this is a first attempt at making it more explicit what the confusing "at least ten times" means. It appears this has been misinterpreted by some as "ten times" meaning that a site needs to provide ten separate warnings/options to extend, when in fact it appears this, like the previous point about adjusting timing, is meant as "ten times the length of time".