Modified the bit on native targets
Took out the bit that the SC applies to the spacing between native elements (I think it that is clear), but addressing @JAWS-test's comments, added an exception for the spacing of targets _inside_ unstyled native elements (i.e. where the vertical spacing of options in a select is determined by the default browser behaviour).