Improving image legends
I improved the ficaption legends of the illustration, some were hard to understand.
I improved the image legends / figcaption texts, and removed the heading Scope and applicability, would have needed a further subheading to separate what are essentialy examples with illustrations. It is odd that there is an actual section Examples further down whithout illustrations, but I do not want to carry through a complete overhaul of the Understanding page at this point. I think it would generally be benefitial to have more subheadings in the longier understanding pages, not just Intent / Benefits / examples, but I am not sure how flexible we are in this respect...