point out that boundaries do not need to contrast
While answering Github issue 856, the group pointed out that boundaries do not need to contrast (unless they are the only way to identify the control). I feel this is a little unclear from the current Boundaries section in the Understanding of 1.4.11.
Therefore the wording was changed to make this more clear.