Adding a clearer ref to 2.1.1
Issue #1397 suggested Suggest adding the word "Pointer" to "2.5.7 Dragging" to match 2.5.1 and 2.52. I hoe the section on distinuishing Dragging movements from (athbased) pointer gestures will satisfy this request.
I have also specified that it is 2.1.1 Keyboard and 2.1.3 Keyboard (No Exception) that require dragging features to be keyboard accessible but that this is not the same as the requirement for single pointer activation required here.