Update text-spacing.html
Removed second paragraph. I tried, but really could not make sense of exactly what the paragraph was trying to convey.
The author responsibility section immediately following touches on the idea of the SC being a baseline which authors are encouraged to exceed.
I have put this in as a separate commit, in case it is felt the paragraph has value, making it easier to keep out of the PR without losing other editorial fixes.