Add clarification that link text is/includes accessible name
in the case of 2.4.4 (which is aimed squarely at AT users), it's essentially link text = accessible name.
in the case of 2.4.9 it's about considering both the "label" for sighted/non-AT users *and* the accessible name for AT users

Note that in this PR I linked directly to the definitions, since currently the XSLT that constructs the understanding pages is unable to pull
definitions for terms that do not appear in the actual SC text - once/if this is fixed, the links should be turned into whatever magic incantation is needed to pull the definition into the page directly and link to it in-page.