Alternative rewording of the normative text for 1.4.2 Audio control
In contrast to https://github.com/w3c/wcag/pull/1824 this rewording accepts that the mechanism does not necessarily have to be implemented by the web page itself, but - per the definition of mechanism - can also be provided by the user agent/system. What this change does implement though is the clarification for the pause/stop clause that the pausing/stopping must be *independent* from any other audio output (which does address @jake-abma's original concern that a system wide general mute/audio off would automatically satisfy the SC, which is not the intended outcome https://github.com/w3c/wcag/issues/1533

This still changes the structure of the SC text, as suggested by @bruce-usab, to a more understandable list format.

If the change in structure is deemed too radical, the addition of "independenlty from any other audio output" could likely be slotted into the current wording/structure of the SC. while a normative change, I don't believe this actually changes the meaning as intended at the time.

Closes https://github.com/w3c/wcag/issues/1533