Reword F23
in line with the proposed clarification for 1.4.2 https://github.com/w3c/wcag/pull/1825 this clarifies that a *mechanism* needs to be available. the mechanism can actually be provided by the UA/system, not necessarily by the Web page. Key here is that it must be possible to turn off the sound *independently* from any other audio output.

Closes https://github.com/w3c/wcag/issues/1543

Only makes sense to merge if https://github.com/w3c/wcag/pull/1825 (or an equivalent/modified version of the PR) is also merged