Update pointer-gestures.html
Added slight changes about understanding. Already has excerpt on "difficulty memorizing" so not too much needed.
Put back a modified version of the prior benefit, that is more tightly constrained.
Unable to add info on double tap, as requested by Gundula (and +1'ed by John Kirkwood) as double taps are clearly allowed as single pointer interaction. As well, OS settings allow a user to significantly alter the timing or interaction to replicate a double tap. So it is not author-dependent.