Changes to 2.5.1 Pointer Gestures understanding (#1569)
* Additional example of a single-pointer alternative to map pinch-zoom multi-pointer gesture

to clarify that this is not only about adding more controls, but rethinking multi-pointer gestures to work for single pointer/finger

* Emphasise that the slider/carousel relies on a fast swipe/flick gesture

swiping on its own could be interpreted as a dragging motion (which is allowed). this hopefully drives home it's more about a very directional and non-continuous flick/swipe

* Remove side effect benefit

As this SC can be passed without adding additional buttons/controls, this benefit is not always present. Having it here may wrongly suggest that the way to pass this always involves actual buttons/controls to be added. This addresses https://github.com/w3c/wcag/issues/747

* Expand on the benefits a bit

cover both path-based aspect (and remind readers that it's not just about touchscreen, but also mouse) and single-pointer aspect

* Update pointer-gestures.html

Added slight changes about understanding. Already has excerpt on "difficulty memorizing" so not too much needed.
Put back a modified version of the prior benefit, that is more tightly constrained.
Unable to add info on double tap, as requested by Gundula (and +1'ed by John Kirkwood) as double taps are clearly allowed as single pointer interaction. As well, OS settings allow a user to significantly alter the timing or interaction to replicate a double tap. So it is not author-dependent.

Co-authored-by: michael-gower <michael.gower@ca.ibm.com>