Adding a section of conforming alternate versions
Adding a paragraph on meeting this SC via conforming alternate versions, taking the input type=number as an example (following discussion in the WCAG issues meeting today). I don't feel very happy about it, partly because it uses input type=number as an example. This feels odd for two reasons: 
1. native input type number would fall under the user agent exception, so not need a conforming alternate version for that reason;
2. input and the arrows generated in some UA are essentially one control so it feels odd treating this control as two separate things, the field being treated as an alterate version to the arrows. Alternate versions would usually be distinct (like a number field below a custom slider that is not keyboard accessible).