Issue 1875 on visible contols
There is an interesting question about scroll bars that only appear on hover and c=focus and buttons that are a the p=bottom of scroll areas. I tried to address that with this PR and also the other questions posed in the SC, but created an FAQ section. We may want to start an FAQ section in our understanding documents where we could answer popular misunderstandings of an SC.