target size of obscured targets
Addressing #1872 asking for clarification regarding dynamic content that obscures other targets
- Requirement does not apply to targets while they are fully or partly obsured by dynamic content
- Requirement does however apply to additional targets brought up by interaction

An issue to be discussed is if this exception equally holds for content that appears not as a result of user interaction but as a result of scripted behaviour of the content (think "Take a survey now" dialogs or cookie banners). Such content will often violate other SCs already (such as 2.4.3 Focus Order and 2.2.1 Timign Ajustable) but I guess both types of dynamic content should be treated the same in terms of the target size requirement.