clarifying name, role, value
closes #2142

In an attempt to add clarity that it's not just that a name, role and value are exposed but the **correct** name role and value are exposed, I have simply added the word "accurate" to two sentences.  This will help imply that it's not just "information" that needs to be exposed to AT, but "accurate information" about what the control represents.

If more modifications are necessary here, please let me know and I can make those changes, or the editors can make changes as they see fit.