add in new paragraph
largely pulling in [Alastair's new paragraph suggestion](https://github.com/w3c/wcag/issues/2142#issuecomment-1029016889), but I modified it a bit to try and use language similar to what ARIA and HTML use (e.g., what the control **represents** dictates its intended semantics and thus the important accessibility information that would be appropriate to convey).

Also, noticed the use of the word "must" in a note in this informative document.  so, changed that to "needs to" to keep the intent.