Addressing issue #2479
Added language to clarify that The language of the success criterion is an OR statement between the requirement for an alternative for time-based media OR an audio description. Therefore, if an audio description cannot be provided because there are insufficient spaces in the dialog to insert the Audio Description, then an alternative for time-based media would be required.