Replace misleading image with clearer example
the current image to illustrate how a simple change of background colour is not in scope of 1.4.11 non-text contrast actually muddies things, because the focused button also has a different text colour and a white halo/shadow around the text.
this example *only* changes the background colour - and, to keep it more consistent with the other examples, it uses the light blue for the non-focused buttons and the dark blue for the focused one.

Closes #2494