Address survey comments (#2682)
* Address survey comments

These changes attempt to address comments from the Task Force regarding the work statement survey.

* Update work-statement.php

* Back out changes for non-web ICT 

Tried, but didn't have consensus on modifying non-web ICT to non-web documents and software. The group seems to prefer a definition on what technologies are covered similar to what exists in the current WCAG2ICT document.

* Update Liaisons & communication sections

Added the GitHub wiki to communications section. Also made modifications to the liaisons bullets and added a proposed editor's note to try to address the 29 Sept. comments.

* Update Timeline section

Make coordination of timeline with regulations/standards a little more generic. Per conversation in 29 September meeting on this section.

* Facilitation section update

Added editors' note to indicate the names will be updated when the position is filled.

* Timeline section - add WCAG2ICT

The document and note needed context by adding "WCAG2ICT".

* add sub-sections

* fix tag

Co-authored-by: Michael Cooper <cooper@w3.org>