List SC 4.1.1 as obsolete
See #2820. Instead of removing the SC, I think we should have marked it as obsolete, with a note explaining not to fail it for modern technologies. I think this causes far fewer down-stream problems. Even if we don't do an errata for WCAG 2.0 and 2.1 (which I think we should), this note makes it clear that 4.1.1 should not be failed there either.