Title case all occurrences of "success criterion"/"success criteria"
This has been the soft editorial style that seems to have been enforced in the last year or so, but not consistently. If this is indeed what we're aiming for, we should likely do it consistently throughout the spec and all related materials, unless the use of the title-cased version has more nuance?

x-ref https://github.com/w3c/wcag2ict/issues/255#issuecomment-1873040201