Improve readability of Understanding 2.5.8 - target size minimum text (#3699)
Closes #3473 

The aim of this PR is to adjust the wording and formatting of the
understanding page for 2.5.8. I've tried to introduce more simplified,
self-contained sentences in order to improve the readability and flow of
the page. This includes:

- Introducing more punctuation for natural breaks
- Removing some repetition of information to reduce the length of
sentences; and
- Introducing unordered lists for blocks of text that discuss more than
two different examples.

Based on online tests, this has improved the readability of the
document.

---------

Co-authored-by: Patrick H. Lauke <redux@splintered.co.uk>
Co-authored-by: Mike Gower <mikegower@gmail.com>
Co-authored-by: Hidde de Vries <160571138+hidde-logius@users.noreply.github.com>
Co-authored-by: DJ Chase <u9000@posteo.mx>