Update F91 to remove unrelated techniques
closes #3788

removes f46, g115 and g63 as they aren't directly related to not marking up table headers correctly.

The issue said all but H43 should be removed, but H51 demonstrates how to build a table with correctly marked up headers, so i see no reason why this shouldn't still be linked.  As it stands, support for the headers attribute is rather lack luster - so pointing to a generic use of setting up a table with correct usage of the th element is worth not removing.