Clean up gh-pages (#3952)
The `gh-pages` branch contains many files/directories that haven't been
actively updated in many years. This PR removes everything that is not
currently relevant to the deploy action on the main branch.

In addition to removing old files, this also updates the following:

- Adds missing style rules to `guidelines.css`
- Redirects `21/guidelines/index.html` to `guidelines/` (like
`22/guidelines/index.html` already does to `guidelines/22/`), rather
than being a slightly older copy of it
- Updates top-level `index.html` to include link to 2.2 guidelines, and
to use newer URL to 2.1 guidelines

(Note that the URLs in the redirect pages mentioned above are hard-coded
to the upstream repo)

This intentionally does _not_ clear outdated files from under
`techniques` and `understanding` (since it seems the existing deploy
process never removes files to clean up what has been removed from the
repository). This can be handled via a deploy script update at a later
point in time.

---------

Co-authored-by: Kevin White <kevin@w3.org>