Change 'non-text graphics' to the normative SC term 'non-text content' (#3158)
The "Testing Principles" section covers the two kinds of non-text
content in scope for this SC: components and graphical objects. However,
the introductory sentence said "non-text graphics" — this could be
construed as limiting the scope of the testing just to the graphics (not
the controls), and possibly further limiting it to just the "non-text"
kind of graphics (if such things exist). So I changed "non-text
graphics" to "non-text content" as in the normative SC.