Harmonise wording for techniques checks/steps (#3915)
Currently, techniques use a mix of wording `If step X...` and `If check
X...`, as well as a few outlier "If test procedure X...". Most prefix
the step/check number with `#`, but some don't. And conceptually, I
think it makes more sense saying that a *check* is true/false, rather
than a *step* being true/false.

This PR tries to harmonise this, settling for `check` and adding missing
`#` (though, it has to be said, this use of the octothorpe is a bit
US-centric, so maybe we want to harmonise to *not* having it?)

---------

Co-authored-by: ljoakley <32885548+ljoakley@users.noreply.github.com>
Co-authored-by: Mike Gower <mikegower@gmail.com>