Clean up mentions of HTML4 and XHTML (#3980)
A lot of HTML4 and XHTML descriptions were removed by #2490, but there
are still some parts left. This PR attempts to clean up those pieces.

## Modify link from HTML4 to HTML

F40 and F44

## Remove mentions of XHTML in techniques

ARIA6, SCR22, C27, F46, F58, F69, F70, F77, F80, G108, G110, G175 and
G204

## Remove mentions of XHTML in understanding

- keyboard
- pronunciation (potentially related
https://github.com/w3c/wcag/issues/2652)

---

However, the following pages are excluded:

[Understanding SC
3.1.1](https://www.w3.org/WAI/WCAG22/Understanding/language-of-page):

```html
<a href="https://www.w3.org/International/techniques/authoring-html#gslang">Internationalization Best Practices: Specifying Language in XHTML &amp; HTML Content</a>
```

This link text is probably inappropriate, but I didn't know what to
change it to, so I left it as it is.

[Understanding
Conformance](https://www.w3.org/WAI/WCAG22/Understanding/conformance):

In the Examples of Conformance Claims section, there are several
mentions of XHTML 1.0. I didn't change them in this PR, but I think it
needs to be rewritten separately.

---------

Co-authored-by: Patrick H. Lauke <redux@splintered.co.uk>