Add top-level index.html for dev only (not GH Pages / W3C site) (#4081)
This is a re-do of #4076, with one change: it only pre-clears the output
folder (`_site`) for W3C site builds, not for GitHub Pages builds. The
deploy script for the latter depends on the output folder _not_ being
cleared, in order to layer the new build on top of existing contents.
(Fortunately this failed loudly the first time upon deploy, so it didn't
affect the contents of the gh-pages branch.)

I tested using my own fork this time to make sure the
GitHub Pages build would complete, and confirmed that
the top-level index.html added in this PR does _not_ output for that build,
leaving the existing top-level index.html in the gh-pages branch intact.