Generate guidelines during Netlify PR preview builds (#4141)
This effectively replaces what the pr-preview bot currently does (and
what #2220 sought to do within the repo). The problem with the
pr-preview bot is its previews don't fully work (e.g. as seen in #4123).

## Changes

- Remove `.pr-preview.json` to stop pr-preview bot from running
- Add logic to `eleventy.config.ts` to copy `guidelines` folder assets
and run through spec-generator (the same thing we use for GH Pages
builds; also the same thing used by the pr-preview bot, IIUC) when
running a standard build through netlify
- Conditionally add guidelines list item to top-level `index.html`
(which is only used for local dev and PR builds)
- Tangentially, rename our github-pages push workflow to be more
distinguishable in the Actions tab