Eleventy build: Support generating WCAG 2.1 docs from main branch (#4007)
## Background

There has long been interest in re-generating updated versions of the
informative docs for WCAG 2.1. Previously, this relied upon a separate
`WCAG-2.1` branch, but the ideal outcome is to be able to build directly
from `main`, altering/pruning content as appropriate.

## Changes

This adds more functionality to the Eleventy-based build system to
support building content targeting WCAG 2.1 in addition to 2.2,
specifically when `WCAG_VERSION=21` is set in environment variables:

- Guidelines are pulled from the web so as to not pick up information
and alterations from newer WCAG versions (e.g. addition of new SC and
removal of 4.1.1)
- Parsing/processing is augmented as necessary to support both source
and pre-processed guidelines
- Acknowledgements are also pulled from the web, pinned to 2.1
- `target-size-enhanced` is output to `target-size` for 2.1
- Detects what techniques only apply to later WCAG versions, and prunes
them from associations
- Avoids emitting pages for techniques and understanding documents that
exclusively pertain to later WCAG versions
- Further updates hard-coded version numbers in titles and level-1
headings, as well as throughout pages (especially `refer-to-wcag`) via
Liquid expressions
- Adds support for setting both `note` and `wcagXY` class on the same
element
- Conditionalizes a couple of 2.2-specific notes that were not within
elements with `wcag22` class
- Unchanged but worth noting: Obsolete technique data includes version
information, so anything marked obsolete as of 2.2 will not contain an
obsolete banner when publishing for 2.1