pull in suggestions
resolved some comments left by @fstrr 

pulled in a version of the suggestions made by @mbgower and @alastc - but i did not delete the text "in regards to the text's intended direction of reading" - as that is meant to tie into the next paragraph.  

if people still have an issue with this, then happy to look at ways to reword.  but i think it's important to make this aspect of the intent clearer