Allow running builds against a custom local guidelines folder (#4182)
This adds logic to `getPrinciples` and `getTermsMap` code flows to allow
explicitly specifying a local `guidelines/index.html` path when building
for a specific version, to be used in place of requesting the latest
published recommendation for the specified `WCAG_VERSION`. This allows
us to directly reuse the build process from the `main` branch for
generating preview builds against `WCAG-2.1`, rather than needing to
copy the entire Eleventy build process to that branch and keep it in
sync.