remove swimlanes and adjust content accordingly
instead of pulling in the edits @fstrr made, i'm removing the initial paragraph that talked about carousels and swimlanes.  The term swimlanes was confusing, and the paragraph wasn't as necessary with the preceding paragraph edits and the fact that carousels are a highlighted example only a couple paragraphs later.

other revisions to the carousel example section per the decision to remove the term swimlanes

see - https://github.com/w3c/wcag/pull/4055#discussion_r1946686423