Comment out 'benefits' section in reflow.html
i'm commenting this out for now.  as mentioned, not every understanding doc has a benefits section.  This benefits section made sense in the old version of this understanding doc - but with the new in brief content, and the heavily revised content of this doc, this benefits section is redundant to other content now in this doc, and otherwise lacking in detail.