Improve wording of first example for path-based gesture
Under Intent, we currently have:

> A **path-based gesture** involves an interaction where not just the endpoints matter, but how the pointer moves between these points.
>  
> If the gesture is only recognised if the user moves in a (mostly) straight line from the start point to the end point, it is an example of a path-based gesture.

That second sentence is confusing.  It is an example of a common characteristic of some path-based gestures.

First example of path-based gesture 

If the gesture is only recognised if the user moves in a (mostly) straight line from the start point to the end point, it is an example of a path-based gesture.