Clarify title of Failure F108 (#1211)
The title "Failure of Success Criterion 2.5.X Dragging due to not
providing a single pointer method for the user to operate a function
that does not require a dragging movement" is ambiguous since the last
part, "that does not require a dragging movement" invites the
interpretation that the single pointer operation alternative has to be
provided for a function that does not use dragging - which is clearly
nonsense.

Simplest solution: Delete "that does not require a dragging movement".

---------

Co-authored-by: Patrick H. Lauke <redux@splintered.co.uk>
Co-authored-by: Mike Gower <mikegower@gmail.com>