Migrate Reflow videos out of repository (#4308)
Resolves #768

Alternative to #4304 that instead migrates videos to W3C's Cloudflare
Stream account. The videos have been marked as downloadable, so they can
be embedded identically to how they were before.

In addition to removing the mp4s from the repo, I've cleaned up a
commented-out gif, and made sure each video has alternative text.

(Note that although the videos are removed, they are still in git
history and thus still consume space; the intent is for this to be the
pattern going forward and to avoid adding more mp4 files to the
repository.)


[Preview](https://deploy-preview-4308--wcag2.netlify.app/understanding/reflow)

Co-authored-by: Mike Gower <mikegower@gmail.com>