Modifed note 1 of 1.2.3 and 1.2.3 Understanding documents
For note 1 of the Understanding document, matched the change to note 3 in the audio description definition by adding "important" in front of "information"
In the same note, swapped out "provided" in favour of "conveyed", in the belief that is more accurate. For example, if a scene shows something blowing up, and there is the sound of an explosion that matches, it is more accurate to say the sound conveys the visual explosion than to say the sound "provides" the visual explosion.