Added "important" to the first note
Two changes to the first note in the informative Understanding document:

- add "additional" to match the first change to the definition note.
- swap out the word "provided" in preference for the synonym "conveyed"

This change is made in the belief that is more accurate. For example, if a scene shows something blowing up, and there is the sound of an explosion that matches, it is more accurate to say the sound conveys the visual explosion than to say the sound "provides" the visual explosion.