Removed 1.2.7 reference from note 1
This change aligns the guidance in the Understanding documents with the SC text for 1.2.7.

The preamble for 1.2.7 states "Where pauses in foreground audio are insufficient to allow audio descriptions to convey the send of the video..."
Based on that preamble, it makes no sense for note 1 to reference 1.2.7, since if all important information is "already conveyed in the audio track", it will never meet the preamble condition.

This logic is confirmed by the fact that note 1 has never existed in the 1.2.7 Understanding document.