Update media-alternative-prerecorded.html
Only one note exists in 1.2.8; however, it contains multiple paragraphs which are enumerated as separate notes in 1.2.3 and 1.2.5.
The first paragraph of 1.2.8, which corresponds to note 1 in 1.2.3 and 1.2.5, is irrelevant to 1.2.8.
By removing this paragraph, 1.2.8 is reduced to coverrng a single concept, which results in an unnumbered note, by convention.