Correct numbers and formulae for focus area calculations
Closes https://github.com/w3c/wcag/issues/4275

Apart from the numbers in the first example being wrong, the actual formulae were wrong as well.

For the rectangle, the simplest calculation is "the larger rectangle (with the extra 2px outline on all sides) minus the base control size"

For the circle, you can't just take the circumference and multiply it by 2. You have to do the same calculation - a circle whose radius is 2px larger, minus the base control circle.

For the rectangle with rounded edges, the calculation comes out very complex - in short, it's 4 rectangles (2 times a rectangle with width minus 2r, multiplied by 2 height; 2 times a rectangle with height minus 2r, multiplied by 2 width) and then a full circle using the previous formula (4 quarters of a circle on each corner)

Bonus: corrected two typos - "accommodation" and "Albania"