Remove two notes
The note about assuming a particular background is contentious at best.

The note about "it is a failure" seems inappropriate in the normative term definition. A failure of what? This sort of note (if it were actually normatively intended) probably should go in the understanding document for a specific SC.