Note #4166 and #4055 in Techniques change log (#4391)
Fixes #4377.

This adds 3 items to the top of the Techniques change log detailing the
1 added technique and 2 removed techniques in #4166. (I have used the
date when #4166 was merged.)

While I was at it, I remembered that #4055 also added new techniques, so
I've added 2 more list items for those.

Initially I had nested/combined list items, but upon further inspection
of the existing items in the change log, I noticed we regularly have
multiple entries with the same date, so I've now flattened the items
added by this PR to one technique per list item.

Tangentially, I also removed punctuation from one line which was
inconsistent with the rest of the top-level list.

@netlify /techniques/

---------

Co-authored-by: Mike Gower <mikegower@gmail.com>