Update media-alternative-prerecorded.html (#4326)
Only one note exists in the 2.1 and 2.2 versions of the 1.2.8
Understanding document; however, it contains multiple paragraphs, which
are enumerated as separate notes in 1.2.3 and 1.2.5 (and in the 1.2.8
Understanding in the 2.0 version). The first paragraph of 1.2.8, which
corresponds to note 1 in 1.2.3 and 1.2.5, is irrelevant to 1.2.8. By
removing this paragraph, 1.2.8's note is reduced to covering a single
concept, which results in an unnumbered note.