Update errata/21.html for previous publications (#4424)
This documents the errata that occurred since each of the 2 previous
(not current) publications, i.e. the publications in September 2023 and
December 2024.

This directly references commit hashes on the WCAG-2.1 branch and their
respective dates (which occurred later), rather than referencing the
original pull requests and the dates they were merged into main.

---------

Co-authored-by: Mike Gower <michael.gower@ca.ibm.com>