JSON: Restore full technique titles within techniquesHtml (#4412)
The implementation of the `techniquesHtml` field in #4301 comes directly
from the already-built Understanding pages. These include logic to
truncate particularly long titles when linking to techniques. This is
similar to what was done in the XSLT build processes both for the
informative docs and `wcag.json`, except that the new build system
includes a visible ellipsis where content was truncated.

For the purposes of JSON output, it makes more sense to give consumers
access to the full title, leaving them free to do what they want with
it.

Since the build process for the JSON output already processes links to
make them absolute, this adds more processing to also fully expand their
titles within the same loop. This impacts relatively few techniques,
many of which are enumerated in #4393.